separating atom and derivation into separate files, using enums for change operation
