+++ lib/src/core/reaction.dart
@@ -1,4 +1,6 @@
-import 'package:mobx/src/core/atom_derivation.dart';
+import 'package:mobx/src/core/atom.dart';
+import 'package:mobx/src/core/context.dart';
+import 'package:mobx/src/core/derivation.dart';
 
 class Reaction implements Derivation {
   Reaction(onInvalidate, {this.name}) {
