+++ lib/src/interceptable.dart
@@ -1,5 +1,5 @@
 import 'package:mobx/src/core/action.dart';
-import 'package:mobx/src/core/atom_derivation.dart';
+import 'package:mobx/src/core/atom.dart';
 
 typedef Interceptor<T> = Function(WillChangeNotification<T>);
