+++ lib/src/listenable.dart
@@ -1,5 +1,5 @@
 import 'package:mobx/src/core/action.dart';
-import 'package:mobx/src/core/atom_derivation.dart';
+import 'package:mobx/src/core/atom.dart';
 
 abstract class Listenable<T> {
   List<Listener<T>> changeListeners;
